+++ src/ExplorerRoutes.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {Route, Routes} from "react-router-dom";
-import LandingPage from "./pages/LandingPage/Dashboard";
+import LandingPage from "./pages/LandingPage/Index";
 import NotFoundPage from "./pages/NotFoundPage";
 import ExplorerLayout from "./pages/layout";
 import TransactionPage from "./pages/Transaction/Index";
