+++ src/pages/Block/Index.tsx
@@ -1,11 +1,11 @@
 import {useParams} from "react-router-dom";
 import {Stack, Grid, Alert} from "@mui/material";
 import React from "react";
-import HeaderSearch from "../layout/Search/Index";
 import BlockTitle from "./Title";
 import BlockTabs from "./Tabs";
 import {useGetBlockByHeight} from "../../api/hooks/useGetBlock";
 import Error from "./Error";
+import PageHeader from "../../components/PageHeader";
 
 export default function BlockPage() {
   const {height} = useParams();
@@ -36,7 +36,7 @@ export default function BlockPage() {
 
   return (
     <Grid container spacing={1}>
-      <HeaderSearch />
+      <PageHeader />
       <Grid item xs={12}>
         <Stack direction="column" spacing={4} marginTop={2}>
           <BlockTitle height={data.block_height} />
