+++ src/pages/Blocks/Index.tsx
@@ -3,6 +3,7 @@ import BlocksTable from "./Table";
 import {useGetMostRecentBlocks} from "../../api/hooks/useGetMostRecentBlocks";
 import {Box, Typography} from "@mui/material";
 import LoadingModal from "../Governance/components/LoadingModal";
+import HeaderSearch from "../layout/Search/Index";
 
 const BLOCKS_COUNT = 30;
 
@@ -13,8 +14,9 @@ export default function BlocksPage() {
     <>
       <LoadingModal open={isLoading} />
       <Box>
+        <HeaderSearch />
         <Typography variant="h5" marginBottom={2}>
-          Recent Blocks
+          Latest Blocks
         </Typography>
         <BlocksTable blocks={recentBlocks} />
       </Box>
