+++ src/pages/layout/FeatureBar.tsx
@@ -44,10 +44,6 @@ export default function FeatureBar() {
     const feature_name = searchParams.get("feature");
     if (feature_name) {
       maybeSetFeature(feature_name);
-    } else {
-      // the "feature" param being null means that it's in "prod"
-      // so set feature to "prod"
-      maybeSetFeature(defaultFeatureName);
     }
   });
